+++ xknx/remote_value/__init__.py
@@ -10,7 +10,6 @@ from .remote_value_climate_mode import (
 from .remote_value_color_rgb import RemoteValueColorRGB
 from .remote_value_color_rgbw import RemoteValueColorRGBW
 from .remote_value_color_xyy import RemoteValueColorXYY
-from .remote_value_control import RemoteValueControl
 from .remote_value_datetime import RemoteValueDateTime
 from .remote_value_dpt_value_1_ucount import RemoteValueDptValue1Ucount
 from .remote_value_raw import RemoteValueRaw
@@ -35,7 +34,6 @@ __all__ = [
     "RemoteValueColorRGB",
     "RemoteValueColorRGBW",
     "RemoteValueColorXYY",
-    "RemoteValueControl",
     "RemoteValueControllerMode",
     "RemoteValueDateTime",
     "RemoteValueDptValue1Ucount",